+++ src/main/clojure/clojure/core/typed.clj
@@ -1085,19 +1085,6 @@ clojure.core.typed/Promise
   [form ty]
   `(ann-form* ~form '~ty))
 
-;(ann unsafe-ann-form* [Any Any -> Any])
-(defn ^:skip-wiki unsafe-ann-form* [form ty]
-  form)
-
-(defmacro unsafe-ann-form 
-  "Assumes the form is well typed and annotates it with the provided
-  type without verifying.
-
-  WARNING: Only use this if you completely understand all the implications and
-  you know what you're doing."
-  [form ty]
-  `(unsafe-ann-form* ~form '~ty))
-
 ;(ann into-array>* [Any Any -> Any])
 (defn ^:skip-wiki
   into-array>*